ratedeathResource changed to float

+++ src/main/java/ir/sharif/aichallenge/server/logic/config/ConstConfigs.java
@@ -19,7 +19,7 @@ public class ConstConfigs {
     public static int MAX_MESSAGE_LENGTH = 10;
 
     public static int GAME_MAXIMUM_TURN_COUNT = 40;
-    public static int RATE_DEATH_RESOURCE = 1;
+    public static float RATE_DEATH_RESOURCE = 1;
 
     public static int MAP_WIDTH = 20;
     public static int MAP_HEIGHT = 10;
